requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
openai==1.3.7
telethon==1.34.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.1
asyncpg==0.29.0
psycopg2>=2.9.0
loguru==0.7.2